GFE WS proof added
TCP is for signaling, TLS is encrypted, and ACK is acknowledged, or when the server acknowledges the PC's request and vice-versa.

It doesn't have anything to do with NSA;
https://forums.geforce.com/default/topic/973629/official-375-70-game-ready-whql-display-driver-feedback-thread-released-10-28-16-/?offset=126

Please stop asking me via eMail if it's 'spying', here is the proof that it's nothing but legitimate traffic. The point is that GFE uses data to build user profiles around games. That type of data collection is not new to nVidia. Origin and Steam collect similar hardware-level data. The data which is shared appears to be purely functional for GFE. NVidia says it may share aggregate-level data with its partners and that it does not share "personally identifiable" information with those partners. There is no way for us to independently validate that, obviously, so we'll have to assume that nVidia is abiding by its own privacy policy and EULA. Steam, as it happens, also has aggregate-level data on its game purchases. You can even view some of this publicly with the correct Chrome plugins. The privacy policy indicates that nVidia can share data between social networks for ad placement via tracking cookies. This is fairly standard, and is done by most websites on the internet -- particularly those which have a product to sell via retail. An example would be when you see Facebook ads that are an exact item you'd been looking at earlier in the month, either through Amazon or elsewhere. We do not see this as a pitchfork-able offense in any greater capacity than we would for any other service.