Disable Telemetry Agent via registry
This does not prevent the setup to install telemetry related files and task, but what it does is that the telemetry agent should never submit something. This is basically the opt-out solution for those who install the "official" untouched versions but don't want to mess with file removal or some scripts.