Updated nvidia bloatware removal
- Renamed to .cmd
- OS check as well as other checks integrated
- MSVCRT, Display.Optimus & NGXCore are now commented out (just remove "REM" if you want to remove them
- Added a param which you can manually execute on older drivers (doesn't work on newer ones) to remove local log files.

I need to re-factor the whole removal process, lots of it is optional since nVidia changed a lot on this. It#s not anymore just to remove Telemetry it's more to avoid "bloatware" and this is typically "as per own needs". So we should re-write it to a PowerShell script to give user a choice what to remove manually during the cleanup session. 

The HOSTS method is typically enough to block telemetry on its own (without removing some components).